requirements.txt
PyQt5>=5.15.0
pyttsx3>=2.90
ffmpeg-python>=0.2.0
googletrans>=3.1.0a0
SpeechRecognition>=3.8.1